front super-items and side images: add the star ratings display (#731)
